energy field no more necessary in Matlab
